+++ plainly-plugin/src/node/collectProject/index.ts
@@ -1,2 +1,1 @@
 export * from './collectProject';
-export * from './settings';
